Using decimal instead of fload for money

+++ Module2/2.Post/Application/Types.fs
@@ -3,4 +3,4 @@
 type Customer = { 
     Id: int
     IsVip: bool
-    Credit: float }
+    Credit: decimal }
