+++ Module3/1.Pre/Application/Types.fs
@@ -5,4 +5,4 @@ open System
 type Customer = { 
     Id: int
     IsVip: bool
-    Credit: float }
+    Credit: decimal }
