+++ Module4/1.Pre/Application/Types.fs
@@ -17,6 +17,6 @@ type Notifications =
 type Customer = {
     Id: int
     IsVip: bool
-    Credit: float<USD>
+    Credit: decimal<USD>
     PersonalDetails: PersonalDetails option
     Notifications: Notifications }
